+++ models/task/src/index.ts
@@ -36,6 +36,7 @@ import attachment from '@anticrm/model-attachment'
 import chunter from '@anticrm/model-chunter'
 import core, { TAttachedDoc, TClass, TDoc, TSpace } from '@anticrm/model-core'
 import view, { actionTemplates as viewTemplates, createAction, template } from '@anticrm/model-view'
+import notification from '@anticrm/notification'
 import { IntlString } from '@anticrm/platform'
 import tags from '@anticrm/tags'
 import {
@@ -373,6 +374,11 @@ export function createModel (builder: Builder): void {
 
   builder.mixin(task.class.Task, core.class.Class, view.mixin.ObjectEditorHeader, {
     editor: task.component.TaskHeader
+  })
+
+  builder.mixin(task.class.Task, core.class.Class, notification.mixin.LastViewAttached, {})
+  builder.mixin(task.class.Task, core.class.Class, notification.mixin.AnotherUserNotifications, {
+    fields: ['assignee']
   })
 
   builder.createDoc(view.class.Viewlet, core.space.Model, {